change raw_padding from double to size_t

+++ src/data_formats.h
@@ -10,7 +10,7 @@ but they are intended to store whole numbers.
 typedef struct data_format
 {
     size_t raw_len;        /* max. possible length of a byte represented in the given data format/representation */
-    double raw_padding;    /* raw, as in it isn't styled/formatted */
+    size_t raw_padding;    /* raw, as in it isn't styled/formatted */
     
     double len;            /* the length (in characters) of a styled data format */
     double group_size;     /* the number of styled data formats in one group */
